Day 15 for Advent of Code 2022
This one was fun! I ended up kinda cheating becaues I couldn't be
bothered to do the math exactly correctly. The key insight
is that the rows will be adjacent to an intersection in the diamond
shape. But there's a problem: I'm lazy! Finding the exact coordinate
that was within/outside the bounds is hard work that I did not
want to do.

So I resolved this by finding the intersection and just adding
all of the rows surrounding it as candidate rows to check.